:green_heart: #30 truncateをAbstractTestCaseクラスに移動
