+++ tests/Feature/AbstractTestCase.php
@@ -6,7 +6,11 @@
 namespace Tests\Feature;
 
 use Tests\TestCase;
+use App\Eloquents\Account;
 use Tests\CreatesApplication;
+use App\Eloquents\AccessToken;
+use App\Eloquents\LoginSession;
+use App\Eloquents\QiitaAccount;
 
 /**
  * Class AbstractTestCase
@@ -15,6 +19,18 @@ use Tests\CreatesApplication;
 abstract class AbstractTestCase extends TestCase
 {
     use CreatesApplication;
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        \DB::statement('SET FOREIGN_KEY_CHECKS=0');
+        Account::truncate();
+        LoginSession::truncate();
+        AccessToken::truncate();
+        QiitaAccount::truncate();
+        \DB::statement('SET FOREIGN_KEY_CHECKS=1');
+    }
 
     public function tearDown()
     {
